requirements.txt
Flask
flask-socketio
eventlet
Pillow
python-dotenv
gunicorn